Clone repo with all history and all tags
Intentional changes from JENKINS-43507 have reduced server load, disc
use, and data transfer by honor the refspec which matches the branch of
the job being built in a multi-branch pipeline, and by not fetching tags.

Unfortunately, this branch contains tests which assume it is operating
with a complete clone of the repository, including all tags and all
branches.

The tests have been fixed on the master branch, but not on this branch.
Rather than fix the tests on this branch (with the risk that creates),
this change modifies the clone to include all history and all tags.